0.4.9 Main Menu Update Part 1 (Bug Fixes)
Main menu is now getting updated with the new UI and this also makes it
easier to develop later.

This also will involve making a toolset we can hopefully sell on the
asset store which will be great to get even a little bit of cashflow

the bug fixes are for things like the inventory not spawning items for
players properly and crafting being a bit missed out on.